Añadidas todas las funciones del storage al servicio y borrado de la carpeta storage
