+++ src/main/kotlin/org/example/fichanewteam/plantilla/service/FileFormat.kt
@@ -1,4 +1,4 @@
-package org.example.fichanewteam.plantilla.storage
+package org.example.fichanewteam.plantilla.service
 
 // Para distinguir los distintos tipos de ficheros
 enum class FileFormat {
