--- src/main/kotlin/org/example/fichanewteam/plantilla/storage/PlantillaStorage.kt
@@ -1,19 +1,0 @@
-package org.example.fichanewteam.plantilla.storage
-
-import com.github.michaelbull.result.Result
-import org.example.fichanewteam.plantilla.error.PlantillaError
-import org.example.fichanewteam.plantilla.models.Plantilla
-import java.io.File
-
-interface PlantillaStorage {
-    fun readFile(file: File, format: FileFormat): List<Plantilla>
-    fun writeFile(file: File, format: FileFormat, personal: List<Plantilla>)
-    fun storageDataJson(file: File, data: List<Plantilla>): Result<Long, PlantillaError>
-    fun deleteAllImages(): Result<Long, PlantillaError>
-    fun loadDataJson(file: File): Result<List<Plantilla>, PlantillaError>
-    fun loadImage(imagenName: String): Result<File, PlantillaError>
-    fun saveImage(fileName: File):Result<File, PlantillaError>
-    fun deleteImage(fileName: String):Result<Unit, PlantillaError>
-    fun exportToZip(zipFile: File, data: List<Plantilla>): Result<File, PlantillaError>
-    fun loadFromZip(unzipFile: File): Result<List<Plantilla>, PlantillaError>
-}
