+++ src/main/kotlin/org/example/fichanewteam/routes/RoutesManager.kt
@@ -104,9 +104,6 @@ object RoutesManager {
     fun initHelpStage() {
             logger.debug { "Cargando vista HELP" }
             val fxmlLoader = FXMLLoader(getResource(View.HELP.fxml))
-        fxmlLoader.controllerFactory = javafx.util.Callback { clazz ->
-            getKoin().get(clazz.kotlin)
-        }
             val root = fxmlLoader.load<Pane>()
             val newScene = Scene(root, 600.0, 400.0)
             Stage().apply {
@@ -140,9 +137,6 @@ object RoutesManager {
 
     fun initPlantillaStage() {
         val fxmlLoader = FXMLLoader(getResource(View.PLANTILLA.fxml))
-        fxmlLoader.controllerFactory = javafx.util.Callback { clazz ->
-            getKoin().get(clazz.kotlin)
-        }
         val root = fxmlLoader.load<Pane>()
         val newScene = Scene(root, 1160.0, 720.0)
         Stage().apply {
